+++ packages/test/src/client/cerbos.test.ts
@@ -609,6 +609,1004 @@ describe("Client", () => {
           });
         });
 
+        describe("withPrincipal", () => {
+          describe("checkResource", () => {
+            it("checks a principal's permissions on a resource", async () => {
+              const result = await clients.default
+                .withPrincipal({
+                  id: "me@example.com",
+                  policyVersion: "1",
+                  scope: "test",
+                  roles: ["USER"],
+                  attr: {
+                    country: {
+                      alpha2: "",
+                      alpha3: "NZL",
+                    },
+                  },
+                })
+                .checkResource({
+                  resource: {
+                    kind: "document",
+                    id: "mine",
+                    policyVersion: "1",
+                    scope: "test",
+                    attr: {
+                      owner: "me@example.com",
+                    },
+                  },
+                  actions: ["view", "edit", "delete"],
+                  auxData: {
+                    jwt: {
+                      token: new UnsecuredJWT({ delete: true }).encode(),
+                    },
+                  },
+                  includeMetadata: true,
+                  requestId: "42",
+                });
+
+              const outputs: OutputResult[] = cerbosVersionIsAtLeast("0.27.0")
+                ? [
+                    {
+                      source: "resource.document.v1#delete",
+                      value: "delete_allowed:me@example.com",
+                    },
+                  ]
+                : [];
+
+              expect(result).toEqual(
+                new CheckResourcesResult({
+                  resource: {
+                    kind: "document",
+                    id: "mine",
+                    policyVersion: "1",
+                    scope: "test",
+                  },
+                  actions: {
+                    view: Effect.ALLOW,
+                    edit: Effect.ALLOW,
+                    delete: Effect.ALLOW,
+                  },
+                  validationErrors: [
+                    {
+                      path: "/country/alpha2",
+                      message: "does not match pattern '[A-Z]{2}'",
+                      source: ValidationErrorSource.PRINCIPAL,
+                    },
+                  ],
+                  metadata: {
+                    actions: {
+                      view: {
+                        matchedPolicy: "resource.document.v1/test",
+                        matchedScope: "test",
+                      },
+                      edit: {
+                        matchedPolicy: "resource.document.v1/test",
+                        matchedScope: "test",
+                      },
+                      delete: {
+                        matchedPolicy: "resource.document.v1/test",
+                        matchedScope: "",
+                      },
+                    },
+                    effectiveDerivedRoles: ["OWNER"],
+                  },
+                  outputs,
+                }),
+              );
+            });
+
+            it("uses the principal's JWT", async () => {
+              const result = await clients.default
+                .withPrincipal(
+                  {
+                    id: "me@example.com",
+                    policyVersion: "1",
+                    scope: "test",
+                    roles: ["USER"],
+                    attr: {
+                      country: {
+                        alpha2: "",
+                        alpha3: "NZL",
+                      },
+                    },
+                  },
+                  {
+                    jwt: {
+                      token: new UnsecuredJWT({ delete: true }).encode(),
+                    },
+                  },
+                )
+                .checkResource({
+                  resource: {
+                    kind: "document",
+                    id: "mine",
+                    policyVersion: "1",
+                    scope: "test",
+                    attr: {
+                      owner: "me@example.com",
+                    },
+                  },
+                  actions: ["view", "edit", "delete"],
+                  includeMetadata: true,
+                  requestId: "42",
+                });
+
+              const outputs: OutputResult[] = cerbosVersionIsAtLeast("0.27.0")
+                ? [
+                    {
+                      source: "resource.document.v1#delete",
+                      value: "delete_allowed:me@example.com",
+                    },
+                  ]
+                : [];
+
+              expect(result).toEqual(
+                new CheckResourcesResult({
+                  resource: {
+                    kind: "document",
+                    id: "mine",
+                    policyVersion: "1",
+                    scope: "test",
+                  },
+                  actions: {
+                    view: Effect.ALLOW,
+                    edit: Effect.ALLOW,
+                    delete: Effect.ALLOW,
+                  },
+                  validationErrors: [
+                    {
+                      path: "/country/alpha2",
+                      message: "does not match pattern '[A-Z]{2}'",
+                      source: ValidationErrorSource.PRINCIPAL,
+                    },
+                  ],
+                  metadata: {
+                    actions: {
+                      view: {
+                        matchedPolicy: "resource.document.v1/test",
+                        matchedScope: "test",
+                      },
+                      edit: {
+                        matchedPolicy: "resource.document.v1/test",
+                        matchedScope: "test",
+                      },
+                      delete: {
+                        matchedPolicy: "resource.document.v1/test",
+                        matchedScope: "",
+                      },
+                    },
+                    effectiveDerivedRoles: ["OWNER"],
+                  },
+                  outputs,
+                }),
+              );
+            });
+
+            it("allows the principal's JWT to be overridden", async () => {
+              const result = await clients.default
+                .withPrincipal(
+                  {
+                    id: "me@example.com",
+                    policyVersion: "1",
+                    scope: "test",
+                    roles: ["USER"],
+                    attr: {
+                      country: {
+                        alpha2: "",
+                        alpha3: "NZL",
+                      },
+                    },
+                  },
+                  {
+                    jwt: {
+                      token: new UnsecuredJWT({ delete: false }).encode(),
+                    },
+                  },
+                )
+                .checkResource({
+                  resource: {
+                    kind: "document",
+                    id: "mine",
+                    policyVersion: "1",
+                    scope: "test",
+                    attr: {
+                      owner: "me@example.com",
+                    },
+                  },
+                  actions: ["view", "edit", "delete"],
+                  auxData: {
+                    jwt: {
+                      token: new UnsecuredJWT({ delete: true }).encode(),
+                    },
+                  },
+                  includeMetadata: true,
+                  requestId: "42",
+                });
+
+              const outputs: OutputResult[] = cerbosVersionIsAtLeast("0.27.0")
+                ? [
+                    {
+                      source: "resource.document.v1#delete",
+                      value: "delete_allowed:me@example.com",
+                    },
+                  ]
+                : [];
+
+              expect(result).toEqual(
+                new CheckResourcesResult({
+                  resource: {
+                    kind: "document",
+                    id: "mine",
+                    policyVersion: "1",
+                    scope: "test",
+                  },
+                  actions: {
+                    view: Effect.ALLOW,
+                    edit: Effect.ALLOW,
+                    delete: Effect.ALLOW,
+                  },
+                  validationErrors: [
+                    {
+                      path: "/country/alpha2",
+                      message: "does not match pattern '[A-Z]{2}'",
+                      source: ValidationErrorSource.PRINCIPAL,
+                    },
+                  ],
+                  metadata: {
+                    actions: {
+                      view: {
+                        matchedPolicy: "resource.document.v1/test",
+                        matchedScope: "test",
+                      },
+                      edit: {
+                        matchedPolicy: "resource.document.v1/test",
+                        matchedScope: "test",
+                      },
+                      delete: {
+                        matchedPolicy: "resource.document.v1/test",
+                        matchedScope: "",
+                      },
+                    },
+                    effectiveDerivedRoles: ["OWNER"],
+                  },
+                  outputs,
+                }),
+              );
+            });
+          });
+
+          describe("checkResources", () => {
+            it("checks a principal's permissions on a set of resources", async () => {
+              const response = await clients.default
+                .withPrincipal({
+                  id: "me@example.com",
+                  policyVersion: "1",
+                  scope: "test",
+                  roles: ["USER"],
+                  attr: {
+                    country: {
+                      alpha2: "",
+                      alpha3: "NZL",
+                    },
+                  },
+                })
+                .checkResources({
+                  resources: [
+                    {
+                      resource: {
+                        kind: "document",
+                        id: "mine",
+                        policyVersion: "1",
+                        scope: "test",
+                        attr: {
+                          owner: "me@example.com",
+                        },
+                      },
+                      actions: ["view", "edit", "delete"],
+                    },
+                    {
+                      resource: {
+                        kind: "document",
+                        id: "theirs",
+                        policyVersion: "1",
+                        scope: "test",
+                        attr: {
+                          owner: "them@example.com",
+                        },
+                      },
+                      actions: ["view", "edit", "delete"],
+                    },
+                    {
+                      resource: {
+                        kind: "document",
+                        id: "invalid",
+                        policyVersion: "1",
+                        scope: "test",
+                        attr: {
+                          owner: 123,
+                        },
+                      },
+                      actions: ["view", "edit", "delete"],
+                    },
+                  ],
+                  auxData: {
+                    jwt: {
+                      token: new UnsecuredJWT({ delete: true }).encode(),
+                    },
+                  },
+                  includeMetadata: true,
+                  requestId: "42",
+                });
+
+              const outputs: OutputResult[] = cerbosVersionIsAtLeast("0.27.0")
+                ? [
+                    {
+                      source: "resource.document.v1#delete",
+                      value: "delete_allowed:me@example.com",
+                    },
+                  ]
+                : [];
+
+              expect(response).toEqual(
+                new CheckResourcesResponse({
+                  requestId: "42",
+                  results: [
+                    new CheckResourcesResult({
+                      resource: {
+                        kind: "document",
+                        id: "mine",
+                        policyVersion: "1",
+                        scope: "test",
+                      },
+                      actions: {
+                        view: Effect.ALLOW,
+                        edit: Effect.ALLOW,
+                        delete: Effect.ALLOW,
+                      },
+                      validationErrors: [
+                        {
+                          path: "/country/alpha2",
+                          message: "does not match pattern '[A-Z]{2}'",
+                          source: ValidationErrorSource.PRINCIPAL,
+                        },
+                      ],
+                      metadata: {
+                        actions: {
+                          view: {
+                            matchedPolicy: "resource.document.v1/test",
+                            matchedScope: "test",
+                          },
+                          edit: {
+                            matchedPolicy: "resource.document.v1/test",
+                            matchedScope: "test",
+                          },
+                          delete: {
+                            matchedPolicy: "resource.document.v1/test",
+                            matchedScope: "",
+                          },
+                        },
+                        effectiveDerivedRoles: ["OWNER"],
+                      },
+                      outputs,
+                    }),
+                    new CheckResourcesResult({
+                      resource: {
+                        kind: "document",
+                        id: "theirs",
+                        policyVersion: "1",
+                        scope: "test",
+                      },
+                      actions: {
+                        view: Effect.ALLOW,
+                        edit: Effect.DENY,
+                        delete: Effect.ALLOW,
+                      },
+                      validationErrors: [
+                        {
+                          path: "/country/alpha2",
+                          message: "does not match pattern '[A-Z]{2}'",
+                          source: ValidationErrorSource.PRINCIPAL,
+                        },
+                      ],
+                      metadata: {
+                        actions: {
+                          view: {
+                            matchedPolicy: "resource.document.v1/test",
+                            matchedScope: "test",
+                          },
+                          edit: {
+                            matchedPolicy: "resource.document.v1/test",
+                            matchedScope: "",
+                          },
+                          delete: {
+                            matchedPolicy: "resource.document.v1/test",
+                            matchedScope: "",
+                          },
+                        },
+                        effectiveDerivedRoles: [],
+                      },
+                      outputs,
+                    }),
+                    new CheckResourcesResult({
+                      resource: {
+                        kind: "document",
+                        id: "invalid",
+                        policyVersion: "1",
+                        scope: "test",
+                      },
+                      actions: {
+                        view: Effect.ALLOW,
+                        edit: Effect.DENY,
+                        delete: Effect.ALLOW,
+                      },
+                      validationErrors: [
+                        {
+                          path: "/country/alpha2",
+                          message: "does not match pattern '[A-Z]{2}'",
+                          source: ValidationErrorSource.PRINCIPAL,
+                        },
+                        {
+                          path: "/owner",
+                          message: "expected string, but got number",
+                          source: ValidationErrorSource.RESOURCE,
+                        },
+                      ],
+                      metadata: {
+                        actions: {
+                          view: {
+                            matchedPolicy: "resource.document.v1/test",
+                            matchedScope: "test",
+                          },
+                          edit: {
+                            matchedPolicy: "resource.document.v1/test",
+                            matchedScope: "",
+                          },
+                          delete: {
+                            matchedPolicy: "resource.document.v1/test",
+                            matchedScope: "",
+                          },
+                        },
+                        effectiveDerivedRoles: [],
+                      },
+                      outputs,
+                    }),
+                  ],
+                }),
+              );
+            });
+
+            it("uses the principal's JWT", async () => {
+              const response = await clients.default
+                .withPrincipal(
+                  {
+                    id: "me@example.com",
+                    policyVersion: "1",
+                    scope: "test",
+                    roles: ["USER"],
+                    attr: {
+                      country: {
+                        alpha2: "",
+                        alpha3: "NZL",
+                      },
+                    },
+                  },
+                  {
+                    jwt: {
+                      token: new UnsecuredJWT({ delete: true }).encode(),
+                    },
+                  },
+                )
+                .checkResources({
+                  resources: [
+                    {
+                      resource: {
+                        kind: "document",
+                        id: "mine",
+                        policyVersion: "1",
+                        scope: "test",
+                        attr: {
+                          owner: "me@example.com",
+                        },
+                      },
+                      actions: ["view", "edit", "delete"],
+                    },
+                    {
+                      resource: {
+                        kind: "document",
+                        id: "theirs",
+                        policyVersion: "1",
+                        scope: "test",
+                        attr: {
+                          owner: "them@example.com",
+                        },
+                      },
+                      actions: ["view", "edit", "delete"],
+                    },
+                    {
+                      resource: {
+                        kind: "document",
+                        id: "invalid",
+                        policyVersion: "1",
+                        scope: "test",
+                        attr: {
+                          owner: 123,
+                        },
+                      },
+                      actions: ["view", "edit", "delete"],
+                    },
+                  ],
+                  includeMetadata: true,
+                  requestId: "42",
+                });
+
+              const outputs: OutputResult[] = cerbosVersionIsAtLeast("0.27.0")
+                ? [
+                    {
+                      source: "resource.document.v1#delete",
+                      value: "delete_allowed:me@example.com",
+                    },
+                  ]
+                : [];
+
+              expect(response).toEqual(
+                new CheckResourcesResponse({
+                  requestId: "42",
+                  results: [
+                    new CheckResourcesResult({
+                      resource: {
+                        kind: "document",
+                        id: "mine",
+                        policyVersion: "1",
+                        scope: "test",
+                      },
+                      actions: {
+                        view: Effect.ALLOW,
+                        edit: Effect.ALLOW,
+                        delete: Effect.ALLOW,
+                      },
+                      validationErrors: [
+                        {
+                          path: "/country/alpha2",
+                          message: "does not match pattern '[A-Z]{2}'",
+                          source: ValidationErrorSource.PRINCIPAL,
+                        },
+                      ],
+                      metadata: {
+                        actions: {
+                          view: {
+                            matchedPolicy: "resource.document.v1/test",
+                            matchedScope: "test",
+                          },
+                          edit: {
+                            matchedPolicy: "resource.document.v1/test",
+                            matchedScope: "test",
+                          },
+                          delete: {
+                            matchedPolicy: "resource.document.v1/test",
+                            matchedScope: "",
+                          },
+                        },
+                        effectiveDerivedRoles: ["OWNER"],
+                      },
+                      outputs,
+                    }),
+                    new CheckResourcesResult({
+                      resource: {
+                        kind: "document",
+                        id: "theirs",
+                        policyVersion: "1",
+                        scope: "test",
+                      },
+                      actions: {
+                        view: Effect.ALLOW,
+                        edit: Effect.DENY,
+                        delete: Effect.ALLOW,
+                      },
+                      validationErrors: [
+                        {
+                          path: "/country/alpha2",
+                          message: "does not match pattern '[A-Z]{2}'",
+                          source: ValidationErrorSource.PRINCIPAL,
+                        },
+                      ],
+                      metadata: {
+                        actions: {
+                          view: {
+                            matchedPolicy: "resource.document.v1/test",
+                            matchedScope: "test",
+                          },
+                          edit: {
+                            matchedPolicy: "resource.document.v1/test",
+                            matchedScope: "",
+                          },
+                          delete: {
+                            matchedPolicy: "resource.document.v1/test",
+                            matchedScope: "",
+                          },
+                        },
+                        effectiveDerivedRoles: [],
+                      },
+                      outputs,
+                    }),
+                    new CheckResourcesResult({
+                      resource: {
+                        kind: "document",
+                        id: "invalid",
+                        policyVersion: "1",
+                        scope: "test",
+                      },
+                      actions: {
+                        view: Effect.ALLOW,
+                        edit: Effect.DENY,
+                        delete: Effect.ALLOW,
+                      },
+                      validationErrors: [
+                        {
+                          path: "/country/alpha2",
+                          message: "does not match pattern '[A-Z]{2}'",
+                          source: ValidationErrorSource.PRINCIPAL,
+                        },
+                        {
+                          path: "/owner",
+                          message: "expected string, but got number",
+                          source: ValidationErrorSource.RESOURCE,
+                        },
+                      ],
+                      metadata: {
+                        actions: {
+                          view: {
+                            matchedPolicy: "resource.document.v1/test",
+                            matchedScope: "test",
+                          },
+                          edit: {
+                            matchedPolicy: "resource.document.v1/test",
+                            matchedScope: "",
+                          },
+                          delete: {
+                            matchedPolicy: "resource.document.v1/test",
+                            matchedScope: "",
+                          },
+                        },
+                        effectiveDerivedRoles: [],
+                      },
+                      outputs,
+                    }),
+                  ],
+                }),
+              );
+            });
+
+            it("allows the principal's JWT to be overridden", async () => {
+              const response = await clients.default
+                .withPrincipal(
+                  {
+                    id: "me@example.com",
+                    policyVersion: "1",
+                    scope: "test",
+                    roles: ["USER"],
+                    attr: {
+                      country: {
+                        alpha2: "",
+                        alpha3: "NZL",
+                      },
+                    },
+                  },
+                  {
+                    jwt: {
+                      token: new UnsecuredJWT({ delete: false }).encode(),
+                    },
+                  },
+                )
+                .checkResources({
+                  resources: [
+                    {
+                      resource: {
+                        kind: "document",
+                        id: "mine",
+                        policyVersion: "1",
+                        scope: "test",
+                        attr: {
+                          owner: "me@example.com",
+                        },
+                      },
+                      actions: ["view", "edit", "delete"],
+                    },
+                    {
+                      resource: {
+                        kind: "document",
+                        id: "theirs",
+                        policyVersion: "1",
+                        scope: "test",
+                        attr: {
+                          owner: "them@example.com",
+                        },
+                      },
+                      actions: ["view", "edit", "delete"],
+                    },
+                    {
+                      resource: {
+                        kind: "document",
+                        id: "invalid",
+                        policyVersion: "1",
+                        scope: "test",
+                        attr: {
+                          owner: 123,
+                        },
+                      },
+                      actions: ["view", "edit", "delete"],
+                    },
+                  ],
+                  auxData: {
+                    jwt: {
+                      token: new UnsecuredJWT({ delete: true }).encode(),
+                    },
+                  },
+                  includeMetadata: true,
+                  requestId: "42",
+                });
+
+              const outputs: OutputResult[] = cerbosVersionIsAtLeast("0.27.0")
+                ? [
+                    {
+                      source: "resource.document.v1#delete",
+                      value: "delete_allowed:me@example.com",
+                    },
+                  ]
+                : [];
+
+              expect(response).toEqual(
+                new CheckResourcesResponse({
+                  requestId: "42",
+                  results: [
+                    new CheckResourcesResult({
+                      resource: {
+                        kind: "document",
+                        id: "mine",
+                        policyVersion: "1",
+                        scope: "test",
+                      },
+                      actions: {
+                        view: Effect.ALLOW,
+                        edit: Effect.ALLOW,
+                        delete: Effect.ALLOW,
+                      },
+                      validationErrors: [
+                        {
+                          path: "/country/alpha2",
+                          message: "does not match pattern '[A-Z]{2}'",
+                          source: ValidationErrorSource.PRINCIPAL,
+                        },
+                      ],
+                      metadata: {
+                        actions: {
+                          view: {
+                            matchedPolicy: "resource.document.v1/test",
+                            matchedScope: "test",
+                          },
+                          edit: {
+                            matchedPolicy: "resource.document.v1/test",
+                            matchedScope: "test",
+                          },
+                          delete: {
+                            matchedPolicy: "resource.document.v1/test",
+                            matchedScope: "",
+                          },
+                        },
+                        effectiveDerivedRoles: ["OWNER"],
+                      },
+                      outputs,
+                    }),
+                    new CheckResourcesResult({
+                      resource: {
+                        kind: "document",
+                        id: "theirs",
+                        policyVersion: "1",
+                        scope: "test",
+                      },
+                      actions: {
+                        view: Effect.ALLOW,
+                        edit: Effect.DENY,
+                        delete: Effect.ALLOW,
+                      },
+                      validationErrors: [
+                        {
+                          path: "/country/alpha2",
+                          message: "does not match pattern '[A-Z]{2}'",
+                          source: ValidationErrorSource.PRINCIPAL,
+                        },
+                      ],
+                      metadata: {
+                        actions: {
+                          view: {
+                            matchedPolicy: "resource.document.v1/test",
+                            matchedScope: "test",
+                          },
+                          edit: {
+                            matchedPolicy: "resource.document.v1/test",
+                            matchedScope: "",
+                          },
+                          delete: {
+                            matchedPolicy: "resource.document.v1/test",
+                            matchedScope: "",
+                          },
+                        },
+                        effectiveDerivedRoles: [],
+                      },
+                      outputs,
+                    }),
+                    new CheckResourcesResult({
+                      resource: {
+                        kind: "document",
+                        id: "invalid",
+                        policyVersion: "1",
+                        scope: "test",
+                      },
+                      actions: {
+                        view: Effect.ALLOW,
+                        edit: Effect.DENY,
+                        delete: Effect.ALLOW,
+                      },
+                      validationErrors: [
+                        {
+                          path: "/country/alpha2",
+                          message: "does not match pattern '[A-Z]{2}'",
+                          source: ValidationErrorSource.PRINCIPAL,
+                        },
+                        {
+                          path: "/owner",
+                          message: "expected string, but got number",
+                          source: ValidationErrorSource.RESOURCE,
+                        },
+                      ],
+                      metadata: {
+                        actions: {
+                          view: {
+                            matchedPolicy: "resource.document.v1/test",
+                            matchedScope: "test",
+                          },
+                          edit: {
+                            matchedPolicy: "resource.document.v1/test",
+                            matchedScope: "",
+                          },
+                          delete: {
+                            matchedPolicy: "resource.document.v1/test",
+                            matchedScope: "",
+                          },
+                        },
+                        effectiveDerivedRoles: [],
+                      },
+                      outputs,
+                    }),
+                  ],
+                }),
+              );
+            });
+          });
+
+          describe("isAllowed", () => {
+            it("checks if a principal is allowed to perform an action on a resource", async () => {
+              const allowed = await clients.default
+                .withPrincipal({
+                  id: "me@example.com",
+                  policyVersion: "1",
+                  scope: "test",
+                  roles: ["USER"],
+                  attr: {
+                    country: {
+                      alpha2: "",
+                      alpha3: "NZL",
+                    },
+                  },
+                })
+                .isAllowed({
+                  resource: {
+                    kind: "document",
+                    id: "mine",
+                    policyVersion: "1",
+                    scope: "test",
+                    attr: {
+                      owner: "me@example.com",
+                    },
+                  },
+                  action: "delete",
+                  auxData: {
+                    jwt: {
+                      token: new UnsecuredJWT({ delete: true }).encode(),
+                    },
+                  },
+                  includeMetadata: true,
+                  requestId: "42",
+                });
+
+              expect(allowed).toBe(true);
+            });
+
+            it("uses the principal's JWT", async () => {
+              const allowed = await clients.default
+                .withPrincipal(
+                  {
+                    id: "me@example.com",
+                    policyVersion: "1",
+                    scope: "test",
+                    roles: ["USER"],
+                    attr: {
+                      country: {
+                        alpha2: "",
+                        alpha3: "NZL",
+                      },
+                    },
+                  },
+                  {
+                    jwt: {
+                      token: new UnsecuredJWT({ delete: true }).encode(),
+                    },
+                  },
+                )
+                .isAllowed({
+                  resource: {
+                    kind: "document",
+                    id: "mine",
+                    policyVersion: "1",
+                    scope: "test",
+                    attr: {
+                      owner: "me@example.com",
+                    },
+                  },
+                  action: "delete",
+                  includeMetadata: true,
+                  requestId: "42",
+                });
+
+              expect(allowed).toBe(true);
+            });
+
+            it("allows the principal's JWT to be overridden", async () => {
+              const allowed = await clients.default
+                .withPrincipal(
+                  {
+                    id: "me@example.com",
+                    policyVersion: "1",
+                    scope: "test",
+                    roles: ["USER"],
+                    attr: {
+                      country: {
+                        alpha2: "",
+                        alpha3: "NZL",
+                      },
+                    },
+                  },
+                  {
+                    jwt: {
+                      token: new UnsecuredJWT({ delete: false }).encode(),
+                    },
+                  },
+                )
+                .isAllowed({
+                  resource: {
+                    kind: "document",
+                    id: "mine",
+                    policyVersion: "1",
+                    scope: "test",
+                    attr: {
+                      owner: "me@example.com",
+                    },
+                  },
+                  action: "delete",
+                  auxData: {
+                    jwt: {
+                      token: new UnsecuredJWT({ delete: true }).encode(),
+                    },
+                  },
+                  includeMetadata: true,
+                  requestId: "42",
+                });
+
+              expect(allowed).toBe(true);
+            });
+          });
+        });
+
         it("handles errors", async () => {
           try {
             await clients.default.checkResources({